# docs/swagger.yaml
definitions:
...
  ErrorResponse:
    properties:
      error:
        properties:
          code:
            enum:
            - FORBIDDEN
            - UNAUTHORIZED
            - INVALID_CREDENTIALS
            - INVALID_REQUEST
            - INTERNAL_ERROR
            - FORM_NOT_FOUND
            - FORM_ALREADY_REJECTED
            - FORM_ALREADY_APPROVED
            - NO_AVAILABLE_EXECUTORS
            - USER_NOT_FOUND
            - USER_NOT_ACTIVE
            - USER_ALREADY_EXISTS
            - DOCUMENT_NOT_FOUND
            example: ERROR_CODE
            type: string
          message:
            type: string
        required:
        - code
        - message
        type: object
    required:
    - error
    type: object
...
  UserResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
      position:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: RESTful API для приложения HRMate. Управление пользователями, авторизацией
    и формами.
  title: HRMate API
  version: "1.0"
paths:
...
  /admin/user:
    get:
      consumes:
      - application/json
      description: Get user based on the requester's role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
  /admin/users:
    get:
      consumes:
      - application/json
      description: |-
        Get users based on the requester's role.
        **Admin**: Returns all users with roles 'employee' and 'hr'
        **HR**: Returns only users with role 'employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users
      tags:
      - Users
  /admin/users/{id}/activate:
    patch:
      consumes:
      - application/json
      description: Change a user's active status to true
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Activate user
      tags:
      - Users
  /admin/users/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Change a user's active status to false
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate user
      tags:
      - Users
...
  /hr/user/{id}:
    get:
      consumes:
      - application/json
      description: Get user based on the requester's role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
  /hr/users:
    get:
      consumes:
      - application/json
      description: |-
        Get users based on the requester's role.
        **Admin**: Returns all users with roles 'employee' and 'hr'
        **HR**: Returns only users with role 'employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users
      tags:
      - Users
...
  /me:
    get:
      consumes:
      - application/json
      description: Returns the profile information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"